# api/api/ai/chat/v1/chat_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v5.29.3
// source: ai/chat/v1/chat.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationChatCreateChatConversation = "/ai.chat.v1.Chat/CreateChatConversation"
const OperationChatDeleteChatConversation = "/ai.chat.v1.Chat/DeleteChatConversation"
const OperationChatDeleteMessage = "/ai.chat.v1.Chat/DeleteMessage"
const OperationChatDeleteUnpinnedChatConversations = "/ai.chat.v1.Chat/DeleteUnpinnedChatConversations"
const OperationChatGetChatConversation = "/ai.chat.v1.Chat/GetChatConversation"
const OperationChatListChatConversationMe = "/ai.chat.v1.Chat/ListChatConversationMe"
const OperationChatListChatConversations = "/ai.chat.v1.Chat/ListChatConversations"
const OperationChatListConversationMessage = "/ai.chat.v1.Chat/ListConversationMessage"
const OperationChatPatchMessage = "/ai.chat.v1.Chat/PatchMessage"
const OperationChatRetryMessage = "/ai.chat.v1.Chat/RetryMessage"
const OperationChatSendMessage = "/ai.chat.v1.Chat/SendMessage"
const OperationChatUpdateChatConversation = "/ai.chat.v1.Chat/UpdateChatConversation"

type ChatHTTPServer interface {
	// CreateChatConversation Conversation 相关接口
	CreateChatConversation(context.Context, *CreateChatConversationRequest) (*GetChatConversationResponse, error)
	DeleteChatConversation(context.Context, *SimpleChatConversationRequest) (*emptypb.Empty, error)
	DeleteMessage(context.Context, *DeleteMessageRequest) (*emptypb.Empty, error)
	DeleteUnpinnedChatConversations(context.Context, *emptypb.Empty) (*emptypb.Empty, error)
	GetChatConversation(context.Context, *SimpleChatConversationRequest) (*GetChatConversationResponse, error)
	ListChatConversationMe(context.Context, *emptypb.Empty) (*GetMultiConversationsResponse, error)
	ListChatConversations(context.Context, *ListConversationRequest) (*ListConversationResponse, error)
	ListConversationMessage(context.Context, *ListConversationMessagesRequest) (*ListConversationMessagesResponse, error)
	PatchMessage(context.Context, *PatchMessageRequest) (*SendMessageResponse, error)
	RetryMessage(context.Context, *RetryMessageRequest) (*SendMessageResponse, error)
	// SendMessage Message 相关接口
	SendMessage(context.Context, *SendMessageRequest) (*SendMessageResponse, error)
	UpdateChatConversation(context.Context, *UpdateChatConversationRequest) (*GetChatConversationResponse, error)
}

func RegisterChatHTTPServer(s *http.Server, srv ChatHTTPServer) {
	r := s.Route("/")
	r.POST("/ai/chat/conversation", _Chat_CreateChatConversation0_HTTP_Handler(srv))
	r.PUT("/ai/chat/conversation/{id}", _Chat_UpdateChatConversation0_HTTP_Handler(srv))
	r.GET("/ai/chat/conversation/list/me", _Chat_ListChatConversationMe0_HTTP_Handler(srv))
	r.GET("/ai/chat/conversation/{id}", _Chat_GetChatConversation1_HTTP_Handler(srv))
	r.DELETE("/ai/chat/conversation/{id}", _Chat_DeleteChatConversation0_HTTP_Handler(srv))
	r.DELETE("/ai/chat/conversation/unpinned", _Chat_DeleteUnpinnedChatConversations0_HTTP_Handler(srv))
	r.POST("/ai/chat/conversation/list", _Chat_ListChatConversations0_HTTP_Handler(srv))
	r.POST("/ai/chat/message/send", _Chat_SendMessage0_HTTP_Handler(srv))
	r.DELETE("/ai/chat/message/{id}", _Chat_DeleteMessage0_HTTP_Handler(srv))
	r.POST("/ai/chat/message/list", _Chat_ListConversationMessage0_HTTP_Handler(srv))
	r.GET("/ai/chat/message/{id}/retry", _Chat_RetryMessage0_HTTP_Handler(srv))
	r.PATCH("/ai/chat/message/{id}", _Chat_PatchMessage0_HTTP_Handler(srv))
}

func _Chat_CreateChatConversation0_HTTP_Handler(srv ChatHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateChatConversationRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationChatCreateChatConversation)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateChatConversation(ctx, req.(*CreateChatConversationRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetChatConversationResponse)
		return ctx.Result(200, reply)
	}
}

func _Chat_UpdateChatConversation0_HTTP_Handler(srv ChatHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateChatConversationRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationChatUpdateChatConversation)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateChatConversation(ctx, req.(*UpdateChatConversationRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetChatConversationResponse)
		return ctx.Result(200, reply)
	}
}

func _Chat_ListChatConversationMe0_HTTP_Handler(srv ChatHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in emptypb.Empty
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationChatListChatConversationMe)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListChatConversationMe(ctx, req.(*emptypb.Empty))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetMultiConversationsResponse)
		return ctx.Result(200, reply)
	}
}

func _Chat_GetChatConversation1_HTTP_Handler(srv ChatHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SimpleChatConversationRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationChatGetChatConversation)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetChatConversation(ctx, req.(*SimpleChatConversationRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetChatConversationResponse)
		return ctx.Result(200, reply)
	}
}

func _Chat_DeleteChatConversation0_HTTP_Handler(srv ChatHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SimpleChatConversationRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationChatDeleteChatConversation)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteChatConversation(ctx, req.(*SimpleChatConversationRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _Chat_DeleteUnpinnedChatConversations0_HTTP_Handler(srv ChatHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in emptypb.Empty
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationChatDeleteUnpinnedChatConversations)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteUnpinnedChatConversations(ctx, req.(*emptypb.Empty))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _Chat_ListChatConversations0_HTTP_Handler(srv ChatHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListConversationRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationChatListChatConversations)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListChatConversations(ctx, req.(*ListConversationRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListConversationResponse)
		return ctx.Result(200, reply)
	}
}

func _Chat_SendMessage0_HTTP_Handler(srv ChatHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SendMessageRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationChatSendMessage)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SendMessage(ctx, req.(*SendMessageRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SendMessageResponse)
		return ctx.Result(200, reply)
	}
}

func _Chat_DeleteMessage0_HTTP_Handler(srv ChatHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteMessageRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationChatDeleteMessage)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteMessage(ctx, req.(*DeleteMessageRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _Chat_ListConversationMessage0_HTTP_Handler(srv ChatHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListConversationMessagesRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationChatListConversationMessage)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListConversationMessage(ctx, req.(*ListConversationMessagesRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListConversationMessagesResponse)
		return ctx.Result(200, reply)
	}
}

func _Chat_RetryMessage0_HTTP_Handler(srv ChatHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RetryMessageRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationChatRetryMessage)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RetryMessage(ctx, req.(*RetryMessageRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SendMessageResponse)
		return ctx.Result(200, reply)
	}
}

func _Chat_PatchMessage0_HTTP_Handler(srv ChatHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in PatchMessageRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationChatPatchMessage)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.PatchMessage(ctx, req.(*PatchMessageRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SendMessageResponse)
		return ctx.Result(200, reply)
	}
}

type ChatHTTPClient interface {
	// CreateChatConversation Conversation 相关接口
	CreateChatConversation(ctx context.Context, req *CreateChatConversationRequest, opts ...http.CallOption) (rsp *GetChatConversationResponse, err error)
	DeleteChatConversation(ctx context.Context, req *SimpleChatConversationRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	DeleteMessage(ctx context.Context, req *DeleteMessageRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	DeleteUnpinnedChatConversations(ctx context.Context, req *emptypb.Empty, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	GetChatConversation(ctx context.Context, req *SimpleChatConversationRequest, opts ...http.CallOption) (rsp *GetChatConversationResponse, err error)
	ListChatConversationMe(ctx context.Context, req *emptypb.Empty, opts ...http.CallOption) (rsp *GetMultiConversationsResponse, err error)
	ListChatConversations(ctx context.Context, req *ListConversationRequest, opts ...http.CallOption) (rsp *ListConversationResponse, err error)
	ListConversationMessage(ctx context.Context, req *ListConversationMessagesRequest, opts ...http.CallOption) (rsp *ListConversationMessagesResponse, err error)
	PatchMessage(ctx context.Context, req *PatchMessageRequest, opts ...http.CallOption) (rsp *SendMessageResponse, err error)
	RetryMessage(ctx context.Context, req *RetryMessageRequest, opts ...http.CallOption) (rsp *SendMessageResponse, err error)
	// SendMessage Message 相关接口
	SendMessage(ctx context.Context, req *SendMessageRequest, opts ...http.CallOption) (rsp *SendMessageResponse, err error)
	UpdateChatConversation(ctx context.Context, req *UpdateChatConversationRequest, opts ...http.CallOption) (rsp *GetChatConversationResponse, err error)
}

type ChatHTTPClientImpl struct {
	cc *http.Client
}

func NewChatHTTPClient(client *http.Client) ChatHTTPClient {
	return &ChatHTTPClientImpl{client}
}

// CreateChatConversation Conversation 相关接口
func (c *ChatHTTPClientImpl) CreateChatConversation(ctx context.Context, in *CreateChatConversationRequest, opts ...http.CallOption) (*GetChatConversationResponse, error) {
	var out GetChatConversationResponse
	pattern := "/ai/chat/conversation"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationChatCreateChatConversation))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ChatHTTPClientImpl) DeleteChatConversation(ctx context.Context, in *SimpleChatConversationRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/ai/chat/conversation/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationChatDeleteChatConversation))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ChatHTTPClientImpl) DeleteMessage(ctx context.Context, in *DeleteMessageRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/ai/chat/message/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationChatDeleteMessage))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ChatHTTPClientImpl) DeleteUnpinnedChatConversations(ctx context.Context, in *emptypb.Empty, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/ai/chat/conversation/unpinned"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationChatDeleteUnpinnedChatConversations))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ChatHTTPClientImpl) GetChatConversation(ctx context.Context, in *SimpleChatConversationRequest, opts ...http.CallOption) (*GetChatConversationResponse, error) {
	var out GetChatConversationResponse
	pattern := "/ai/chat/conversation/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationChatGetChatConversation))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ChatHTTPClientImpl) ListChatConversationMe(ctx context.Context, in *emptypb.Empty, opts ...http.CallOption) (*GetMultiConversationsResponse, error) {
	var out GetMultiConversationsResponse
	pattern := "/ai/chat/conversation/list/me"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationChatListChatConversationMe))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ChatHTTPClientImpl) ListChatConversations(ctx context.Context, in *ListConversationRequest, opts ...http.CallOption) (*ListConversationResponse, error) {
	var out ListConversationResponse
	pattern := "/ai/chat/conversation/list"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationChatListChatConversations))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ChatHTTPClientImpl) ListConversationMessage(ctx context.Context, in *ListConversationMessagesRequest, opts ...http.CallOption) (*ListConversationMessagesResponse, error) {
	var out ListConversationMessagesResponse
	pattern := "/ai/chat/message/list"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationChatListConversationMessage))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ChatHTTPClientImpl) PatchMessage(ctx context.Context, in *PatchMessageRequest, opts ...http.CallOption) (*SendMessageResponse, error) {
	var out SendMessageResponse
	pattern := "/ai/chat/message/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationChatPatchMessage))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PATCH", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ChatHTTPClientImpl) RetryMessage(ctx context.Context, in *RetryMessageRequest, opts ...http.CallOption) (*SendMessageResponse, error) {
	var out SendMessageResponse
	pattern := "/ai/chat/message/{id}/retry"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationChatRetryMessage))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// SendMessage Message 相关接口
func (c *ChatHTTPClientImpl) SendMessage(ctx context.Context, in *SendMessageRequest, opts ...http.CallOption) (*SendMessageResponse, error) {
	var out SendMessageResponse
	pattern := "/ai/chat/message/send"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationChatSendMessage))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ChatHTTPClientImpl) UpdateChatConversation(ctx context.Context, in *UpdateChatConversationRequest, opts ...http.CallOption) (*GetChatConversationResponse, error) {
	var out GetChatConversationResponse
	pattern := "/ai/chat/conversation/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationChatUpdateChatConversation))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}