# api/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /ai/chat/conversation:
        post:
            tags:
                - Chat
            description: Conversation 相关接口
            operationId: Chat_CreateChatConversation
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ai.chat.v1.CreateChatConversationRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ai.chat.v1.GetChatConversationResponse'
    /ai/chat/conversation/list:
        post:
            tags:
                - Chat
            operationId: Chat_ListChatConversations
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ai.chat.v1.ListConversationRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ai.chat.v1.ListConversationResponse'
...
    /ai/chat/conversation/{id}:
        get:
            tags:
                - Chat
            operationId: Chat_GetChatConversation
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ai.chat.v1.GetChatConversationResponse'
        put:
            tags:
                - Chat
            operationId: Chat_UpdateChatConversation
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ai.chat.v1.UpdateChatConversationRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ai.chat.v1.GetChatConversationResponse'
        delete:
            tags:
                - Chat
            operationId: Chat_DeleteChatConversation
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content: {}
    /ai/chat/message/list:
        post:
            tags:
                - Chat
            operationId: Chat_ListConversationMessage
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ai.chat.v1.ListConversationMessagesRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ai.chat.v1.ListConversationMessagesResponse'
    /ai/chat/message/send:
        post:
            tags:
                - Chat
            description: Message 相关接口
            operationId: Chat_SendMessage
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ai.chat.v1.SendMessageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ai.chat.v1.SendMessageResponse'
    /ai/chat/message/{id}:
        delete:
            tags:
                - Chat
            operationId: Chat_DeleteMessage
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: cascade
                  in: query
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content: {}
        patch:
            tags:
                - Chat
            operationId: Chat_PatchMessage
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ai.chat.v1.PatchMessageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ai.chat.v1.SendMessageResponse'
    /ai/chat/message/{id}/retry:
        get:
            tags:
                - Chat
            operationId: Chat_RetryMessage
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: useContext
                  in: query
                  schema:
                    type: boolean
                - name: useSearch
                  in: query
                  schema:
                    type: boolean
                - name: modelId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ai.chat.v1.SendMessageResponse'
...
components:
    schemas:
...
        ai.chat.v1.CreateChatConversationRequest:
            type: object
            properties:
                roleId:
                    type: string
                knowledgeId:
                    type: string
                modelId:
                    type: string
        ai.chat.v1.GetChatConversationResponse:
            type: object
            properties:
                id:
                    type: string
                title:
                    type: string
                pinned:
                    type: boolean
                systemMessage:
                    type: string
                model:
                    type: string
                modelId:
                    type: string
                temperature:
                    type: number
                    format: double
                maxTokens:
                    type: string
                maxContexts:
                    type: string
                createdAt:
                    type: string
                    format: date-time
                roleAvatar:
                    type: string
                    description: 关联role信息
                roleName:
                    type: string
                messageCount:
                    type: integer
                    format: int32
...
        ai.chat.v1.KnowledgeSegment:
            type: object
            properties:
                id:
                    type: string
                content:
                    type: string
                documentId:
                    type: string
                documentName:
                    type: string
        ai.chat.v1.ListConversationMessagesRequest:
            type: object
            properties:
                pagination:
                    $ref: '#/components/schemas/common.v1.PaginationArgs'
                conversationId:
                    type: string
        ai.chat.v1.ListConversationMessagesResponse:
            type: object
            properties:
                messages:
                    type: array
                    items:
                        $ref: '#/components/schemas/ai.chat.v1.MessageRecord'
                pagination:
                    $ref: '#/components/schemas/common.v1.PaginationResults'
        ai.chat.v1.ListConversationRequest:
            type: object
            properties:
                pagination:
                    $ref: '#/components/schemas/common.v1.PaginationArgs'
                title:
                    type: string
                start:
                    type: string
                    format: date-time
                end:
                    type: string
                    format: date-time
        ai.chat.v1.ListConversationResponse:
            type: object
            properties:
                conversations:
                    type: array
                    items:
                        $ref: '#/components/schemas/ai.chat.v1.GetChatConversationResponse'
                pagination:
                    $ref: '#/components/schemas/common.v1.PaginationResults'
        ai.chat.v1.MessageRecord:
            type: object
            properties:
                id:
                    type: string
                type:
                    type: string
                content:
                    type: string
                reasonContent:
                    type: string
                segments:
                    type: array
                    items:
                        $ref: '#/components/schemas/ai.chat.v1.KnowledgeSegment'
                webPages:
                    type: array
                    items:
                        $ref: '#/components/schemas/ai.chat.v1.WebPage'
                createdAt:
                    type: string
                    format: date-time
                conversationId:
                    type: string
                replyId:
                    type: string
        ai.chat.v1.PatchMessageRequest:
            type: object
            properties:
                id:
                    type: string
                content:
                    type: string
                useContext:
                    type: boolean
                useSearch:
                    type: boolean
                modelId:
                    type: string
        ai.chat.v1.SendMessageRequest:
            type: object
            properties:
                conversationId:
                    type: string
                content:
                    type: string
                useContext:
                    type: boolean
                useSearch:
                    type: boolean
                attachmentUrls:
                    type: array
                    items:
                        type: string
                modelId:
                    type: string
        ai.chat.v1.SendMessageResponse:
            type: object
            properties:
                send:
                    $ref: '#/components/schemas/ai.chat.v1.MessageRecord'
                receive:
                    $ref: '#/components/schemas/ai.chat.v1.MessageRecord'
                done:
                    type: boolean
        ai.chat.v1.UpdateChatConversationRequest:
            type: object
            properties:
                id:
                    type: string
                title:
                    type: string
                pinned:
                    type: boolean
                systemMessage:
                    type: string
                modelId:
                    type: string
                temperature:
                    type: number
                    format: double
                maxTokens:
                    type: string
                maxContexts:
                    type: string
        ai.chat.v1.WebPage:
            type: object
            properties:
                name:
                    type: string
                icon:
                    type: string
                title:
                    type: string
                url:
                    type: string
                snippet:
                    type: string
                summary:
                    type: string
...
        common.v1.PaginationArgs:
            type: object
            properties:
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
                orderBy:
                    type: string
                orderDirection:
                    type: string
        common.v1.PaginationResults:
            type: object
            properties:
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
                totalItems:
                    type: integer
                    format: int32
...